db connect done
